+++ geo/CHANGES.md
@@ -4,6 +4,7 @@
 
 - Move `PreparedGeometry` into a new `indexed` module intended to provide index-backed geometries. `relate::PreparedGeometry` has been deprecated.
 - Use an interval tree for faster (Multi)Point in MultiPolygon checks
+- LOF algorithm efficiency improvements due to caching kth distance
 
 ## 0.31.0 - 2025-09-01